Refactor code for better unit testing
